add: support for registering editor sessions with file paths and workspace URI configurations; update related utilities and tests
