+++ packages/ace-linters/src/language-provider.ts
@@ -1,17 +1,15 @@
 import {Ace} from "ace-code";
 import "./types/ace-extension";
 
-import {FormattingOptions} from "vscode-languageserver-protocol";
 import {CommonConverter} from "./type-converters/common-converters";
-import {ComboDocumentIdentifier, IMessageController} from "./types/message-controller-interface";
+import {IMessageController} from "./types/message-controller-interface";
 import {MessageController} from "./message-controller";
 import {
-    fromAceDelta, fromAnnotations, fromDocumentHighlights,
+    fromAnnotations,
     fromPoint,
     fromRange, fromSignatureHelp,
-    toAnnotations,
     toCompletionItem,
-    toCompletions, toInlineCompletions, toMarkerGroupItem,
+    toCompletions, toInlineCompletions,
     toRange, toResolvedCompletion,
     toTooltip
 } from "./type-converters/lsp/lsp-converters";
@@ -21,28 +19,23 @@ import showdown from "showdown";
 import {createWorker} from "./cdn-worker";
 import {SignatureTooltip} from "./components/signature-tooltip";
 import {
-    AceRangeData, CodeActionsByService,
+    CodeActionsByService,
     ProviderOptions,
     ServiceFeatures,
     ServiceOptions,
-    ServiceOptionsMap, ServiceStruct,
+    ServiceOptionsMap, ServiceStruct, SessionInitialConfig,
     SupportedServices,
     Tooltip
 } from "./types/language-service";
-import {MarkerGroup} from "./ace/marker_group";
 import {AceRange} from "./ace/range-singleton";
 import {HoverTooltip} from "./ace/hover-tooltip";
-import {
-    DecodedSemanticTokens,
-    mergeTokens,
-    parseSemanticTokens
-} from "./type-converters/lsp/semantic-tokens";
 import {LightbulbWidget} from "./components/lightbulb";
 import {AceVirtualRenderer} from "./ace/renderer-singleton";
 import {AceEditor} from "./ace/editor-singleton";
 import {setStyles} from "./misc/styles";
 import {convertToUri} from "./utils";
 import {createInlineCompleterAdapter} from "./ace/inline_autocomplete";
+import {SessionLanguageProvider} from "./session-language-provider";
 
 export class LanguageProvider {
     activeEditor: Ace.Editor;
@@ -71,7 +64,7 @@ export class LanguageProvider {
     }
 
     /**
-     *  Creates LanguageProvider using our transport protocol with ability to register different services on same
+     *  Creates LanguageProvider using our transport protocol with the ability to register different services on the same
      *  webworker
      * @param {Worker} worker
      * @param {ProviderOptions} options
@@ -174,16 +167,14 @@ export class LanguageProvider {
      * be joined with the workspace URI.
      *
      * @param session The Ace edit session to update with the file path.
-     * @param filePath The file path to set for the session.
-     * @param [joinWorkspaceURI] Optional parameter to determine if the file path should be joined with the workspace URI.
-     * Defaults to false if not provided.
+     * @param config config to set
      */
-    setSessionFilePath(session: Ace.EditSession, filePath: string, joinWorkspaceURI?: boolean) {
-        this.$getSessionLanguageProvider(session)?.setFilePath(filePath, joinWorkspaceURI);
+    setSessionFilePath(session: Ace.EditSession, config: SessionInitialConfig) {
+        this.$getSessionLanguageProvider(session)?.setFilePath(config.filePath, config.joinWorkspaceURI);
     }
 
-    private $registerSession = (session: Ace.EditSession, editor: Ace.Editor) => {
-        this.$sessionLanguageProviders[session["id"]] ??= new SessionLanguageProvider(this, session, editor, this.$messageController);
+    private $registerSession = (session: Ace.EditSession, editor: Ace.Editor, config?: SessionInitialConfig) => {
+        this.$sessionLanguageProviders[session["id"]] ??= new SessionLanguageProvider(this, session, editor, this.$messageController, config);
     }
 
     private $getSessionLanguageProvider(session: Ace.EditSession): SessionLanguageProvider {
@@ -196,17 +187,24 @@ export class LanguageProvider {
     }
 
     /**
-     * Registers an Ace editor instance with the language provider.
-     * @param editor - The Ace editor instance to register.
+     * Registers an Ace editor instance along with the session's configuration settings.
+     *
+     * @param editor - The Ace editor instance to be registered.
+     * @param [config] - Configuration options for the session.
      */
-    registerEditor(editor: Ace.Editor) {
+    registerEditor(editor: Ace.Editor, config?: SessionInitialConfig) {
         if (!this.editors.includes(editor))
             this.$registerEditor(editor);
-        this.$registerSession(editor.session, editor);
+        this.$registerSession(editor.session, editor, config);
     }
 
     codeActionCallback: (codeActions: CodeActionsByService[]) => void;
 
+    /**
+     * Sets a callback function that will be triggered with an array of code actions grouped by service.
+     *
+     * @param {function} callback - A function that receives an array of code actions, categorized by service, as its argument.
+     */
     setCodeActionCallback(callback: (codeActions: CodeActionsByService[]) => void) {
         this.codeActionCallback = callback;
     }
@@ -401,6 +399,13 @@ export class LanguageProvider {
         }
     }
 
+    /**
+     * Sets global options for the specified service.
+     *
+     * @param serviceName - The name of the service for which to set global options.
+     * @param options - The options to set for the specified service.
+     * @param {boolean} [merge=false] - Indicates whether to merge the provided options with the existing options. Defaults to false.
+     */
     setGlobalOptions<T extends keyof ServiceOptionsMap>(serviceName: T & string, options: ServiceOptionsMap[T], merge = false) {
         this.$messageController.setGlobalOptions(serviceName, options, merge);
     }
@@ -422,11 +427,24 @@ export class LanguageProvider {
         this.$messageController.setWorkspace(this.workspaceUri);
     }
 
+    /**
+     * Sets the options for a specified editor session.
+     *
+     * @param session - The Ace editor session to configure.
+     * @param options - The configuration options to be applied to the session.
+     */
     setSessionOptions<OptionsType extends ServiceOptions>(session: Ace.EditSession, options: OptionsType) {
         let sessionLanguageProvider = this.$getSessionLanguageProvider(session);
         sessionLanguageProvider.setOptions(options);
     }
 
+    /**
+     * Configures the specified features for a given service.
+     *
+     * @param {SupportedServices} serviceName - The name of the service for which features are being configured.
+     * @param {ServiceFeatures} features - The features to be configured for the given service.
+     * @return {void} Does not return a value.
+     */
     configureServiceFeatures(serviceName: SupportedServices, features: ServiceFeatures) {
         this.$messageController.configureFeatures(serviceName, features);
     }
@@ -617,362 +635,5 @@ export class LanguageProvider {
                 error: e
             }, serviceName);
         }
-    }
-}
-
-class SessionLanguageProvider {
-    session: Ace.EditSession;
-    documentUri: string;
-    private $messageController: IMessageController;
-    private $deltaQueue: Ace.Delta[] | null;
-    private $isConnected = false;
-    private $options?: ServiceOptions;
-    private $filePath: string;
-    private $servicesCapabilities?: { [serviceName: string]: lsp.ServerCapabilities };
-    private $requestsQueue: Function[] = [];
-
-    state: {
-        occurrenceMarkers: MarkerGroup | null,
-        diagnosticMarkers: MarkerGroup | null
-    } = {
-        occurrenceMarkers: null,
-        diagnosticMarkers: null
-    }
-
-    private extensions = {
-        "typescript": "ts",
-        "javascript": "js"
-    }
-    editor: Ace.Editor;
-
-    private semanticTokensLegend?: lsp.SemanticTokensLegend;
-    private $provider: LanguageProvider;
-
-    /**
-     * Constructs a new instance of the `SessionLanguageProvider` class.
-     *
-     * @param provider - The `LanguageProvider` instance.
-     * @param session - The Ace editor session.
-     * @param editor - The Ace editor instance.
-     * @param messageController - The `IMessageController` instance for handling messages.
-     */
-    constructor(provider: LanguageProvider, session: Ace.EditSession, editor: Ace.Editor, messageController: IMessageController) {
-        this.$provider = provider;
-        this.$messageController = messageController;
-        this.session = session;
-        this.editor = editor;
-
-        session.doc.version = 1;
-        session.doc.on("change", this.$changeListener, true);
-        this.addSemanticTokenSupport(session); //TODO: ?
-        session.on("changeMode", this.$changeMode);
-        if (this.$provider.options.functionality!.semanticTokens) {
-            session.on("changeScrollTop", () => this.getSemanticTokens());
-        }
-
-        this.$init();
-    }
-
-    enqueueIfNotConnected(callback: () => void) {
-        if (!this.$isConnected) {
-            this.$requestsQueue.push(callback);
-        } else {
-            callback();
-        }
-    }
-
-    get comboDocumentIdentifier(): ComboDocumentIdentifier {
-        return {
-            documentUri: this.documentUri,
-            sessionId: this.session["id"]
-        };
-    }
-
-    /**
-     * @param filePath
-     * @param joinWorkspaceURI
-     */
-    setFilePath(filePath: string, joinWorkspaceURI?: boolean) {
-        this.enqueueIfNotConnected(() => {
-            this.session.doc.version++;
-            this.$filePath = filePath;
-            const previousComboId = this.comboDocumentIdentifier;
-            this.initDocumentUri(true, joinWorkspaceURI);
-            this.$messageController.renameDocument(previousComboId, this.comboDocumentIdentifier.documentUri, this.session.doc.version);
-        })
-    };
-
-    private $init() {
-        this.initDocumentUri();
-        this.$messageController.init(this.comboDocumentIdentifier, this.session.doc, this.$mode, this.$options, this.$connected);
-    }
-
-    addSemanticTokenSupport(session: Ace.EditSession) {
-        let bgTokenizer = session.bgTokenizer;
-        session.setSemanticTokens = (tokens: DecodedSemanticTokens | undefined) => {
-            bgTokenizer.semanticTokens = tokens;
-        }
-
-        bgTokenizer.$tokenizeRow = (row: number) => {
-            var line = bgTokenizer.doc.getLine(row);
-            var state = bgTokenizer.states[row - 1];
-            var data = bgTokenizer.tokenizer.getLineTokens(line, state, row);
-
-            if (bgTokenizer.states[row] + "" !== data.state + "") {
-                bgTokenizer.states[row] = data.state;
-                bgTokenizer.lines[row + 1] = null;
-                if (bgTokenizer.currentLine > row + 1)
-                    bgTokenizer.currentLine = row + 1;
-            } else if (bgTokenizer.currentLine == row) {
-                bgTokenizer.currentLine = row + 1;
-            }
-
-            if (bgTokenizer.semanticTokens) {
-                let decodedTokens = bgTokenizer.semanticTokens.getByRow(row);
-                if (decodedTokens) {
-                    data.tokens = mergeTokens(data.tokens, decodedTokens);
-                }
-            }
-
-            return bgTokenizer.lines[row] = data.tokens;
-        }
-    }
-
-    private $connected = (capabilities: { [serviceName: string]: lsp.ServerCapabilities }) => {
-        this.$isConnected = true;
-
-        this.setServerCapabilities(capabilities);
-
-        this.$requestsQueue.forEach((requestCallback) => requestCallback());
-        this.$requestsQueue = [];
-
-        if (this.$deltaQueue)
-            this.$sendDeltaQueue();
-        if (this.$options)
-            this.setOptions(this.$options);
-    }
-
-    private $changeMode = () => {
-        this.enqueueIfNotConnected(() => {
-            this.$deltaQueue = [];
-
-            this.session.clearAnnotations();
-            if (this.state.diagnosticMarkers) {
-                this.state.diagnosticMarkers.setMarkers([]);
-            }
-
-            this.session.setSemanticTokens(undefined); //clear all semantic tokens
-            let newVersion = this.session.doc.version++;
-            this.$messageController.changeMode(this.comboDocumentIdentifier, this.session.getValue(), newVersion, this.$mode, this.setServerCapabilities);
-        });
-    };
-
-    setServerCapabilities = (capabilities: { [serviceName: string]: lsp.ServerCapabilities }) => {
-        if (!capabilities)
-            return;
-        this.$servicesCapabilities = {...capabilities};
-
-        let hasTriggerChars = Object.values(capabilities).some((capability) => capability?.completionProvider?.triggerCharacters);
-
-        if (hasTriggerChars || this.$provider.options.functionality?.completion && this.$provider.options.functionality?.completion.lspCompleterOptions?.triggerCharacters) {
-            let completer = this.editor.completers.find((completer) => completer.id === "lspCompleters");
-            if (completer) {
-                let allTriggerCharacters: string[] = [];
-                Object.values(capabilities).forEach((capability) => {
-                    if (capability?.completionProvider?.triggerCharacters) {
-                        allTriggerCharacters.push(...capability.completionProvider.triggerCharacters);
-                    }
-                });
-
-                allTriggerCharacters = [...new Set(allTriggerCharacters)];
-
-                const triggerCharacterOptions = (typeof this.$provider.options.functionality?.completion == "object") ? this.$provider.options.functionality.completion.lspCompleterOptions?.triggerCharacters : undefined;
-                if (triggerCharacterOptions) {
-                    const removeChars: string[] = Array.isArray(triggerCharacterOptions.remove)
-                        ? triggerCharacterOptions.remove
-                        : [];
-                    const addChars: string[] = Array.isArray(triggerCharacterOptions.add)
-                        ? triggerCharacterOptions.add
-                        : [];
-                    completer.triggerCharacters = allTriggerCharacters.filter(
-                        (char: string) => !removeChars.includes(char)
-                    );
-                    addChars.forEach((char: string) => {
-                        if (!completer!.triggerCharacters!.includes(char)) {
-                            completer!.triggerCharacters!.push(char);
-                        }
-                    });
-                } else {
-                    completer.triggerCharacters = allTriggerCharacters;
-                }
-            }
-        }
-
-        let hasSemanticTokensProvider = Object.values(capabilities).some((capability) => {
-            if (capability?.semanticTokensProvider) {
-                this.semanticTokensLegend = capability.semanticTokensProvider.legend;
-                return true;
-            }
-        });
-        if (hasSemanticTokensProvider) {
-            this.getSemanticTokens();
-        }
-        //TODO: we should restrict range formatting if any of services is only has full format capabilities
-        //or we shoudl use service with full format capability instead of range one's
-    }
-
-    private initDocumentUri(isRename = false, joinWorkspaceURI = false) {
-        let filePath = this.$filePath ?? this.session["id"] + "." + this.$extension;
-        if (isRename) {
-            delete this.$provider.$urisToSessionsIds[this.documentUri];
-        }
-        this.documentUri = convertToUri(filePath, joinWorkspaceURI, this.$provider.workspaceUri);
-        this.$provider.$urisToSessionsIds[this.documentUri] = this.session["id"];
-    }
-
-    private get $extension() {
-        let mode = this.$mode.replace("ace/mode/", "");
-        return this.extensions[mode] ?? mode;
-    }
-
-    private get $mode(): string {
-        return this.session["$modeId"];
-    }
-
-    private get $format(): FormattingOptions {
-        return {
-            tabSize: this.session.getTabSize(),
-            insertSpaces: this.session.getUseSoftTabs()
-        }
-    }
-
-    private $changeListener = (delta) => {
-        this.session.doc.version++;
-        if (!this.$deltaQueue) {
-            this.$deltaQueue = [];
-            setTimeout(() => this.$sendDeltaQueue(() => {
-                this.getSemanticTokens();
-            }), 0);
-        }
-        this.$deltaQueue.push(delta);
-    }
-
-    $sendDeltaQueue = (callback?) => {
-        let deltas = this.$deltaQueue;
-        if (!deltas) return callback && callback();
-        this.$deltaQueue = null;
-        if (deltas.length)
-            this.$messageController.change(this.comboDocumentIdentifier, deltas.map((delta) =>
-                fromAceDelta(delta, this.session.doc.getNewLineCharacter())), this.session.doc, callback);
-    };
-
-    $showAnnotations = (diagnostics: lsp.Diagnostic[]) => {
-        if (!diagnostics) {
-            return;
-        }
-
-        let annotations = toAnnotations(diagnostics);
-        this.session.clearAnnotations();
-        if (annotations && annotations.length > 0) {
-            this.session.setAnnotations(annotations);
-        }
-        if (!this.state.diagnosticMarkers) {
-            this.state.diagnosticMarkers = new MarkerGroup(this.session);
-        }
-        this.state.diagnosticMarkers.setMarkers(diagnostics?.map((el) => toMarkerGroupItem(CommonConverter.toRange(toRange(el.range)), "language_highlight_error", el.message)));
-    }
-
-    setOptions<OptionsType extends ServiceOptions>(options: OptionsType) {
-        if (!this.$isConnected) {
-            this.$options = options;
-            return;
-        }
-        this.$messageController.changeOptions(this.comboDocumentIdentifier, options);
-    }
-
-    validate = () => {
-        this.$messageController.doValidation(this.comboDocumentIdentifier, this.$showAnnotations);
-    }
-
-    format = () => {
-        let selectionRanges = this.session.getSelection().getAllRanges();
-        let $format = this.$format;
-        let aceRangeDatas = selectionRanges as AceRangeData[];
-        if (!selectionRanges || selectionRanges[0].isEmpty()) {
-            let row = this.session.getLength();
-            let column = this.session.getLine(row).length - 1;
-            aceRangeDatas =
-                [{
-                    start: {
-                        row: 0, column: 0
-                    },
-                    end: {
-                        row: row, column: column
-                    }
-                }];
-        }
-        for (let range of aceRangeDatas) {
-            this.$messageController.format(this.comboDocumentIdentifier, fromRange(range), $format, this.applyEdits);
-        }
-    }
-
-    applyEdits = (edits: lsp.TextEdit[]) => {
-        edits ??= [];
-        for (let edit of edits.reverse()) {
-            this.session.replace(<Ace.Range>toRange(edit.range), edit.newText);
-        }
-    }
-
-    getSemanticTokens() {
-        if (!this.$provider.options.functionality!.semanticTokens)
-            return;
-        //TODO: improve this 
-        let lastRow = this.editor.renderer.getLastVisibleRow();
-        let visibleRange: AceRangeData = {
-            start: {
-                row: this.editor.renderer.getFirstVisibleRow(),
-                column: 0
-            },
-            end: {
-                row: lastRow + 1,
-                column: this.session.getLine(lastRow).length
-            }
-        }
-
-        this.$messageController.getSemanticTokens(this.comboDocumentIdentifier, fromRange(visibleRange), (tokens) => {
-                if (!tokens) {
-                    return;
-                }
-                let decodedTokens = parseSemanticTokens(tokens.data, this.semanticTokensLegend!.tokenTypes, this.semanticTokensLegend!.tokenModifiers);
-                this.session.setSemanticTokens(decodedTokens);
-                let bgTokenizer = this.session.bgTokenizer;
-
-                //@ts-ignore
-                bgTokenizer.running = setTimeout(() => {
-                    if (bgTokenizer?.semanticTokens?.tokens && bgTokenizer?.semanticTokens?.tokens.length > 0) {
-                        let startRow: number = bgTokenizer?.semanticTokens?.tokens[0].row;
-                        bgTokenizer.currentLine = startRow;
-                        bgTokenizer.lines = bgTokenizer.lines.slice(0, startRow - 1);
-                    } else {
-                        bgTokenizer.currentLine = 0;
-                        bgTokenizer.lines = [];
-                    }
-                    bgTokenizer.$worker();
-                }, 20);
-            }
-        );
-    }
-
-    $applyDocumentHighlight = (documentHighlights: lsp.DocumentHighlight[]) => {
-        if (!this.state.occurrenceMarkers) {
-            this.state.occurrenceMarkers = new MarkerGroup(this.session);
-        }
-        if (documentHighlights) { //some servers return null, which contradicts spec
-            this.state.occurrenceMarkers.setMarkers(fromDocumentHighlights(documentHighlights));
-        }
-    };
-
-    closeDocument(callback?) {
-        this.$messageController.closeDocument(this.comboDocumentIdentifier, callback);
     }
 }
