+++ packages/ace-linters/src/types/language-service.ts
@@ -263,6 +263,18 @@ export interface ProviderOptions {
     }
 }
 
+export interface SessionInitialConfig {
+    /**
+     * The associated file path for the session
+     */
+    filePath: string,
+    /**
+     * Determines if the editor should join the workspace URI
+     * @default `false`
+     */
+    joinWorkspaceURI?: boolean
+}
+
 export type ServiceFeatures = {
     [feature in SupportedFeatures]?: boolean;
 };
@@ -295,7 +307,9 @@ export interface BaseConfig {
     className?: string,
     features?: ServiceFeatures,
     module: () => any,
-    id?: string
+    id?: string,
+    /** Whether to encode URIs when processing language service operations */
+    encodeUris?: boolean
 }
 
 interface WebWorkerConnection {
