+++ packages/ace-linters/tests/unit/language-provider.tests.ts
@@ -32,7 +32,7 @@ describe('LanguageProvider tests', () => {
 
     before((done) => {
         editorEl = document.createElement('div');
-        
+
         editor = ace.edit(editorEl, {
             value: `
     <!DOCTYPE html>
@@ -80,7 +80,8 @@ describe('LanguageProvider tests', () => {
                 completionResolve: true,
                 format: true,
                 documentHighlights: false,
-                signatureHelp: false
+                signatureHelp: false,
+                codeActions: false
             }
         });
         languageProvider.registerEditor(editor);
@@ -114,7 +115,7 @@ describe('LanguageProvider tests', () => {
                 hover: false,
                 diagnostics: false,
             });
-            
+
             languageProvider.doHover(editor.session, {row: 2, column: 2}, hover => {
                 expect(hover).to.be.undefined;
                 done();
@@ -154,6 +155,179 @@ describe('LanguageProvider tests', () => {
         editor.session.on("change", changeListener);
 
         languageProvider.format();
+    });
+
+    describe('registerEditor with file paths - worker side verification', function () {
+        let testProvider: LanguageProvider;
+        let testClient: MockWorker, testCtx: MockWorker;
+        let testManager: ServiceManager;
+
+        beforeEach(() => {
+            testClient = new MockWorker();
+            testCtx = new MockWorker();
+            testClient.setEmitter(testCtx);
+            testCtx.setEmitter(testClient);
+
+            testManager = new ServiceManager(testCtx);
+            testManager.registerService("html", {
+                features: {completion: true, completionResolve: true, diagnostics: true, format: true, hover: true},
+                module: () => import("../../src/services/html/html-service"),
+                className: "HtmlService",
+                modes: "html"
+            });
+        });
+
+        it('should register editor with custom file path and verify worker service document URI', (done) => {
+            testProvider = LanguageProvider.create(testClient, { functionality: {
+                    codeActions: false
+                }});
+
+            let testEditor = ace.edit(document.createElement('div'), {
+                value: '<div>test content</div>',
+                mode: new HtmlMode()
+            });
+
+            // Wait for the service to be connected and document to be added
+            setTimeout(() => {
+                testProvider.registerEditor(testEditor, {
+                    filePath: 'custom/path/test.html'
+                });
+
+                setTimeout(() => {
+                    const expectedDocumentUri = 'file:///custom/path/test.html';
+                    const htmlService = testManager.getServicesInstances(expectedDocumentUri)[0];
+
+                    expect(htmlService).to.exist;
+                    expect(htmlService.documents[expectedDocumentUri]).to.exist;
+                    done();
+                }, 100);
+            }, 50);
+        });
+
+        it('should register editor with workspace URI and verify worker service document URI', (done) => {
+            testProvider = LanguageProvider.create(testClient, {
+                workspacePath: '/workspace/project',
+                functionality: {
+                    codeActions: false
+                }
+            });
+
+            let testEditor = ace.edit(document.createElement('div'), {
+                value: '<div>workspace test</div>',
+                mode: new HtmlMode()
+            });
+
+            setTimeout(() => {
+                testProvider.registerEditor(testEditor, {
+                    filePath: 'src/components/test.html',
+                    joinWorkspaceURI: true,
+                });
+
+                setTimeout(() => {
+                    const expectedDocumentUri = 'file:///workspace/project/src/components/test.html';
+                    const htmlService = testManager.getServicesInstances(expectedDocumentUri)[0];
+
+                    expect(htmlService).to.exist;
+                    expect(htmlService.documents[expectedDocumentUri]).to.exist;
+                    done();
+                }, 100);
+            }, 50);
+        });
+
+        it('should register editor without joinWorkspaceURI and verify worker service document URI', (done) => {
+            testProvider = LanguageProvider.create(testClient, {
+                workspacePath: '/workspace/project',
+                functionality: {
+                    codeActions: false
+                }
+            });
+
+            let testEditor = ace.edit(document.createElement('div'), {
+                value: '<div>no join test</div>',
+                mode: new HtmlMode()
+            });
+
+            setTimeout(() => {
+                testProvider.registerEditor(testEditor, {
+                    filePath: '/absolute/path/test.html',
+                    joinWorkspaceURI: false
+                });
+
+                setTimeout(() => {
+                    const expectedDocumentUri = 'file:///absolute/path/test.html';
+                    const htmlService = testManager.getServicesInstances(expectedDocumentUri)[0];
+
+                    expect(htmlService).to.exist;
+                    expect(htmlService.documents[expectedDocumentUri]).to.exist;
+                    done();
+                }, 100);
+            }, 50);
+        });
+
+        it('should register editor without config and verify default document URI', (done) => {
+            testProvider = LanguageProvider.create(testClient, {
+                functionality: {
+                    codeActions: false
+                }
+            });
+
+            let testEditor = ace.edit(document.createElement('div'), {
+                value: '<div>default test</div>',
+                mode: new HtmlMode()
+            });
+
+            setTimeout(() => {
+                testProvider.registerEditor(testEditor);
+
+                setTimeout(() => {
+                    // Should use session ID with .html extension
+                    const expectedPattern = `file:///${testEditor.session["id"]}.html`;
+                    const htmlService = testManager.getServicesInstances(expectedPattern)[0];
+
+                    expect(htmlService).to.exist;
+                    expect(htmlService.documents[expectedPattern]).to.exist;
+                    done();
+                }, 100);
+            }, 50);
+        });
+
+        it('should handle setSessionFilePath after registration', (done) => {
+            testProvider = LanguageProvider.create(testClient, {
+                workspacePath: '/workspace/project',
+            });
+
+            let testEditor = ace.edit(document.createElement('div'), {
+                value: '<div>dynamic path test</div>',
+                mode: new HtmlMode()
+            });
+
+            setTimeout(() => {
+                // First register with default
+                testProvider.registerEditor(testEditor);
+
+                setTimeout(() => {
+                    // Then change the file path
+                    testProvider.setSessionFilePath(testEditor.session, {
+                        filePath: 'dynamic/new-path.html',
+                        joinWorkspaceURI: true
+                    });
+
+                    setTimeout(() => {
+                        const expectedDocumentUri = 'file:///workspace/project/dynamic/new-path.html';
+                        const htmlService = testManager.getServicesInstances(expectedDocumentUri)[0];
+
+                        expect(htmlService).to.exist;
+                        expect(htmlService.documents[expectedDocumentUri]).to.exist;
+                        done();
+                    }, 100);
+                }, 100);
+            }, 50);
+        });
+
+        afterEach(() => {
+            testClient.removeAllListeners();
+            testCtx.removeAllListeners();
+        });
     });
 
     describe('json functionality', function () {
@@ -243,12 +417,12 @@ describe('LanguageProvider tests', () => {
                     expect(completions[0].caption).to.equal("property");
                 })
 
-               /* it('do resolve', function (done) {
-                    languageProvider.doResolve(completions[0], function (completionItem) {
-                        expect(completionItem.documentation).to.equal("basic property");
-                        done();
-                    })
-                })*/
+                /* it('do resolve', function (done) {
+                     languageProvider.doResolve(completions[0], function (completionItem) {
+                         expect(completionItem.documentation).to.equal("basic property");
+                         done();
+                     })
+                 })*/
             });
 
         });
@@ -256,19 +430,19 @@ describe('LanguageProvider tests', () => {
 
     it('should remove document from linked services', (done) => {
         let $sessionIDToMode = manager["$sessionIDToMode"];
-        let name = "file:///" + editor.session["id"] +".html";
+        let name = "file:///" + editor.session["id"] + ".html";
         let htmlService = manager.getServicesInstances(name)[0];
-        
+
         expect(htmlService.documents[name]).exist;
         expect($sessionIDToMode[name]).exist;
-        
+
         languageProvider.closeDocument(editor.session, () => {
-            
+
             expect(htmlService.documents[name]).not.exist;
             expect($sessionIDToMode[name]).not.exist;
             done()
         });
-        
+
     })
 
     after(() => {
